Removed $more_link_text from the_content.
Starting from scratch this is the right move. Not quite sure the impact on themes that have already translated this. Adding this as a theme setting now will, I think, override their translation. It would be a quick fix though, one change in theme settings. And it would mean all new users enjoy this setting. Let me know.

Signed-off-by: Andrew Misplon <amisplon@gmail.com>