Correct issue with page meta info.
We inserted page metadata to help with Google Authorship requirements. We then hid this. The way we hid this causes post metadata to be hidden on pages where the Post Loop widget is being used. I've corrected. Could be done more elegantly but I think this works. Ideally this needs a second test. Looks good on my side.

Signed-off-by: Andrew Misplon <amisplon@gmail.com>